fix: revert bingAds changes
