--- __tests__/integrations/BingAds/__mocks__/data.js
@@ -1,17 +1,0 @@
-const event = 'Add to Cart';
-const query = 'Sport T-Shirt';
-const products = [
-  {
-    product_id: '123',
-    sku: 'G-14',
-    price: 14.99,
-    quantity: 2,
-  },
-  {
-    product_id: '345',
-    sku: 'G-32',
-    price: 3.99,
-  },
-];
-
-export { event, query, products };
