+++ src/integrations/BingAds/browser.js
@@ -1,9 +1,6 @@
 import logger from '../../utils/logUtil';
 import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
-import { buildCommonPayload, buildEcommPayload, EXCLUSION_KEYS } from './utils';
-import { removeUndefinedAndNullValues } from '../../utils/commonUtils';
-import { extractCustomFields } from '../../utils/utils';
 
 class BingAds {
   constructor(config, analytics, destinationInfo) {
@@ -12,8 +9,8 @@ class BingAds {
     }
     this.tagID = config.tagID;
     this.name = NAME;
-    this.areTransformationsConnected = destinationInfo?.areTransformationsConnected;
-    this.destinationId = destinationInfo?.destinationId;
+    this.areTransformationsConnected = destinationInfo && destinationInfo.areTransformationsConnected;
+    this.destinationId = destinationInfo && destinationInfo.destinationId;
     this.uniqueId = `bing${this.tagID}`;
   }
 
@@ -28,7 +25,7 @@ class BingAds {
           const o = {
             ti: this.tagID,
           };
-          (o.q = w[u]), (w[u] = new UET(o));
+          (o.q = w[u]), (w.UET && (w[u] = new UET(o)));
         }),
         (n = d.createElement(t)),
         (n.src = r),
@@ -37,7 +34,7 @@ class BingAds {
         (n.onload = n.onreadystatechange =
           function () {
             const s = this.readyState;
-            (s && s !== 'loaded' && s !== 'complete' && typeof w['UET'] === 'function') ||
+            (s && s !== 'loaded' && s !== 'complete') ||
               (f(), (n.onload = n.onreadystatechange = null));
           }),
         (i = d.getElementsByTagName(t)[0]),
@@ -47,15 +44,18 @@ class BingAds {
   /* eslint-enable */
 
   init = () => {
-    this.loadBingadsScript();
+    this.loadBingadsScript();  
     logger.debug('===in init BingAds===');
   };
 
   isLoaded = () => {
     logger.debug('in BingAds isLoaded');
-    return (
-      !!window.UET && !!window[this.uniqueId] && window[this.uniqueId].push !== Array.prototype.push
-    );
+    if(typeof window.UET !== 'function') {
+      logger.debug('BingAds: UET class is yet to be loaded. Retrying.');
+    } else {
+      logger.debug('BingAds: UET class is successfully loaded');
+    }
+    return (!!window.UET && !!window[this.uniqueId] && window[this.uniqueId].push !== Array.prototype.push);
   };
 
   isReady = () => {
@@ -67,33 +67,34 @@ class BingAds {
     Visit here(for details Parameter details): https://help.ads.microsoft.com/#apex/3/en/53056/2
     Under: What data does UET collect once I install it on my website?
     Updated syntax doc ref - https://help.ads.microsoft.com/#apex/ads/en/56916/2
-    Ecomm parameters ref - https://help.ads.microsoft.com/#apex/ads/en/60118/-1
   */
 
   track = (rudderElement) => {
-    const { type, properties } = rudderElement.message;
-    const { eventAction } = properties;
-    const eventToSend = eventAction || type;
+    const { type, properties, event } = rudderElement.message;
+    const { category, currency, value, revenue, total } = properties;
+    const eventToSend = type;
     if (!eventToSend) {
       logger.error('Event type not present');
       return;
     }
-
-    let payload = {
-      ...buildCommonPayload(rudderElement.message),
-      ...buildEcommPayload(rudderElement.message),
+    const payload = {
+      event_label: event,
     };
-
-    // We can pass unmapped UET parameters through custom properties
-    const customProperties = extractCustomFields(
-      rudderElement.message,
-      {},
-      ['properties'],
-      EXCLUSION_KEYS,
-    );
-
-    payload = { ...payload, ...customProperties };
-    payload = removeUndefinedAndNullValues(payload);
+    if (category) {
+      payload.event_category = category;
+    }
+    if (currency) {
+      payload.currency = currency;
+    }
+    if (value) {
+      payload.revenue_value = value;
+    }
+    if (revenue) {
+      payload.revenue_value = revenue;
+    }
+    if (total) {
+      payload.revenue_value = total;
+    }
     window[this.uniqueId].push('event', eventToSend, payload);
   };
